Type mismatch in Portfolio.setPositionId(). Was looking for argument of type int, needed long

+++ src/main/java/net/toddsarratt/GaussTrader/Position.java
@@ -87,7 +87,7 @@ public class Position {
     public long getPositionId() {
         return positionId;
     }
-    void setPositionId(int positionId ) {
+    void setPositionId(long positionId ) {
         this.positionId = positionId;
     }
     void setOpen(boolean open) {
